refactor(scripts): Move scripts so they can be used by latest ORT
This applies the idea of [1] and adapts to [2].

[1]: https://github.com/oss-review-toolkit/ort/pull/6961
[2]: https://github.com/oss-review-toolkit/ort/pull/6965
