--- evaluator-rules/gradle/libs.versions.toml
@@ -1,5 +1,0 @@
-[versions]
-ort = "7df497b47b"
-
-[libraries]
-ortEvaluator = { module = "com.github.oss-review-toolkit.ort:evaluator", version.ref = "ort" }
