--- notifications/build.gradle.kts
@@ -1,56 +1,0 @@
-/*
- * Copyright (C) 2022 The ORT Project Authors (see <https://github.com/oss-review-toolkit/ort-config/blob/main/NOTICE>)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-val ortJitPackGroup: String by project
-val ortJitPackVersion: String by project
-
-plugins {
-    `java-library`
-}
-
-repositories {
-    mavenCentral()
-
-    exclusiveContent {
-        forRepository {
-            maven("https://jitpack.io")
-        }
-
-        filter {
-            includeGroup(libs.ortNotifier.get().module.group)
-        }
-    }
-
-    // Need to repeat several custom repository definitions of other submodules here, see
-    // https://github.com/gradle/gradle/issues/4106.
-    exclusiveContent {
-        forRepository {
-            maven("https://packages.atlassian.com/maven-external")
-        }
-
-        filter {
-            includeGroupByRegex("com\\.atlassian\\..*")
-            includeVersionByRegex("log4j", "log4j", ".*-atlassian-.*")
-        }
-    }
-}
-
-dependencies {
-    compileOnly(libs.ortNotifier)
-}
