--- notifications/gradle/libs.versions.toml
@@ -1,5 +1,0 @@
-[versions]
-ort = "7df497b47b"
-
-[libraries]
-ortNotifier = { module = "com.github.oss-review-toolkit.ort:notifier", version.ref = "ort" }
