+++ reporter.how-to-fix-text-provider.kts
